refactor: use relative app compat consumers
