+++ rustfs/src/app/context/compat.rs
@@ -17,11 +17,11 @@ use super::handles::{
     default_bucket_metadata_interface, default_endpoints_interface, default_kms_runtime_interface,
     default_server_config_interface, default_tier_config_interface,
 };
-use crate::app::context::storage_compat::ECStore;
-use crate::app::context::storage_compat::EndpointServerPools;
-use crate::app::context::storage_compat::TierConfigMgr;
-use crate::app::context::storage_compat::metadata_sys::BucketMetadataSys;
-use crate::app::context::storage_compat::new_object_layer_fn;
+use super::storage_compat::ECStore;
+use super::storage_compat::EndpointServerPools;
+use super::storage_compat::TierConfigMgr;
+use super::storage_compat::metadata_sys::BucketMetadataSys;
+use super::storage_compat::new_object_layer_fn;
 #[cfg(test)]
 use crate::config::RustFSBufferConfig;
 use rustfs_config::server_config::Config;
@@ -119,6 +119,10 @@ fn resolve_buffer_config_with(
 
 #[cfg(test)]
 mod tests {
+    use super::super::storage_compat::Endpoint;
+    use super::super::storage_compat::init_local_disks;
+    use super::super::storage_compat::new_object_layer_fn;
+    use super::super::storage_compat::{Endpoints, PoolEndpoints};
     use super::*;
     use crate::app::context::global::AppContextTestInterfaces;
     use crate::app::context::handles::{default_notify_interface, default_region_interface};
@@ -126,10 +130,6 @@ mod tests {
         BucketMetadataInterface, BufferConfigInterface, EndpointsInterface, IamInterface, KmsInterface, KmsRuntimeInterface,
         ServerConfigInterface, TierConfigInterface,
     };
-    use crate::app::context::storage_compat::Endpoint;
-    use crate::app::context::storage_compat::init_local_disks;
-    use crate::app::context::storage_compat::new_object_layer_fn;
-    use crate::app::context::storage_compat::{Endpoints, PoolEndpoints};
     use crate::config::{RustFSBufferConfig, WorkloadProfile};
     use rustfs_iam::{store::object::ObjectStore, sys::IamSys};
     use std::path::PathBuf;
