+++ rustfs/src/app/context/handles.rs
@@ -16,10 +16,10 @@ use super::interfaces::{
     BucketMetadataInterface, BufferConfigInterface, EndpointsInterface, IamInterface, KmsInterface, KmsRuntimeInterface,
     NotifyInterface, RegionInterface, ServerConfigInterface, TierConfigInterface,
 };
-use crate::app::context::storage_compat::EndpointServerPools;
-use crate::app::context::storage_compat::TierConfigMgr;
-use crate::app::context::storage_compat::metadata_sys::{BucketMetadataSys, get_global_bucket_metadata_sys};
-use crate::app::context::storage_compat::{get_global_endpoints_opt, get_global_region, get_global_tier_config_mgr};
+use super::storage_compat::EndpointServerPools;
+use super::storage_compat::TierConfigMgr;
+use super::storage_compat::metadata_sys::{BucketMetadataSys, get_global_bucket_metadata_sys};
+use super::storage_compat::{get_global_endpoints_opt, get_global_region, get_global_tier_config_mgr};
 use crate::config::{RustFSBufferConfig, get_global_buffer_config};
 use async_trait::async_trait;
 use rustfs_config::server_config::Config;
