+++ rustfs/src/server/readiness.rs
@@ -839,7 +839,7 @@ mod tests {
 
     #[test]
     fn aggregate_lock_quorum_status_requires_each_set_to_meet_quorum() {
-        use crate::server::storage_compat::{Endpoint, EndpointServerPools, Endpoints, PoolEndpoints};
+        use super::super::storage_compat::{Endpoint, EndpointServerPools, Endpoints, PoolEndpoints};
 
         let endpoints = vec![
             Endpoint {
@@ -896,7 +896,7 @@ mod tests {
 
     #[test]
     fn aggregate_lock_quorum_status_fails_when_any_set_loses_quorum() {
-        use crate::server::storage_compat::{Endpoint, EndpointServerPools, Endpoints, PoolEndpoints};
+        use super::super::storage_compat::{Endpoint, EndpointServerPools, Endpoints, PoolEndpoints};
 
         let endpoints = vec![
             Endpoint {
